refactor: LottoNumberGenerator를 인터페이스로 구현
