+++ src/main/java/lotto/domain/lottonumbergenerator/LottoNumberAutoGenerator.java
@@ -1,20 +1,29 @@
-package lotto.domain;
+package lotto.domain.lottonumbergenerator;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import lotto.domain.LottoNumber;
 
-public class LottoNumberGenerator {
+public class LottoNumberAutoGenerator implements LottoNumberGenerator {
 
     private static final int MINIMUM_RANGE = 1;
     private static final int MAXIMUM_RANGE = 45;
 
     private final List<LottoNumber> lottoNumbers;
 
-    public LottoNumberGenerator() {
+    public LottoNumberAutoGenerator() {
         lottoNumbers = createLottoNumbers();
+    }
+
+    @Override
+    public List<LottoNumber> getLottoNumbers(int count) {
+        shuffleNumbers();
+        List<LottoNumber> numbers = lottoNumbers.subList(0, count);
+        Collections.sort(numbers);
+        return new ArrayList<>(numbers);
     }
 
     private List<LottoNumber> createLottoNumbers() {
@@ -25,12 +34,5 @@ public class LottoNumberGenerator {
 
     private void shuffleNumbers() {
         Collections.shuffle(lottoNumbers);
-    }
-
-    public List<LottoNumber> getLottoNumbers(int count) {
-        shuffleNumbers();
-        List<LottoNumber> numbers = lottoNumbers.subList(0, count);
-        Collections.sort(numbers);
-        return new ArrayList<>(numbers);
     }
 }
