+++ src/main/java/lotto/domain/lottonumbergenerator/LottoNumberGenerator.java
@@ -1,0 +1,9 @@
+package lotto.domain.lottonumbergenerator;
+
+import java.util.List;
+import lotto.domain.LottoNumber;
+
+public interface LottoNumberGenerator {
+
+    List<LottoNumber> getLottoNumbers(int count);
+}
