+++ src/test/java/lottoTest/domain/lottonumbergenerator/LottoNumberAutoGeneratorTest.java
@@ -1,0 +1,16 @@
+package lottoTest.domain.lottonumbergenerator;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import lotto.domain.lottonumbergenerator.LottoNumberAutoGenerator;
+import org.junit.jupiter.api.Test;
+
+@SuppressWarnings("NonAsciiCharacters")
+class LottoNumberAutoGeneratorTest {
+
+    @Test
+    void 로또번호를_6개_생성하는_기능() {
+        LottoNumberAutoGenerator lottoNumberAutoGenerator = new LottoNumberAutoGenerator();
+        assertThat(lottoNumberAutoGenerator.getLottoNumbers(6).size()).isEqualTo(6);
+    }
+}
